Migrate to pyproject.toml and update project structure (#281)
* commit

* train: get version from toml instead of __init__

* fixup

* revert black to clarify pr

* Remove setup.py

* Test new build

* try again

* gitignore

* get version a different way

* Try again

* Try again

* tweak buildscript

* catch ffmpeg errors
