+++ lightning_pose/api/model_config.py
