+++ lightning_pose/cli/commands/predict.py
@@ -8,7 +8,7 @@ from typing import TYPE_CHECKING
 from .. import types
 
 if TYPE_CHECKING:
-    from lightning_pose.model import Model
+    from lightning_pose.api.model import Model
 
 
 def register_parser(subparsers):
@@ -76,7 +76,7 @@ def register_parser(subparsers):
 def handle(args):
     """Handle the predict command."""
     # Delay this import because it's slow.
-    from lightning_pose.model import Model
+    from lightning_pose.api.model import Model
 
     model = Model.from_dir2(args.model_dir, hydra_overrides=args.overrides)
     input_paths = [Path(p) for p in args.input_path]
