+++ lightning_pose/cli/friendly.py
