+++ lightning_pose/cli/main.py
@@ -37,6 +37,7 @@ def main():
 
     # Run any migrations (from lightning_pose.migrations).
     from lightning_pose.migrations.migrations import run_migrations
+
     run_migrations()
 
     # Execute the command
