--- setup.cfg
@@ -1,22 +1,0 @@
-[flake8]
-max-line-length = 99
-ignore = F821, W503
-extend-ignore = E203
-exclude =
-    .git,
-    __pycache__,
-    __init__.py,
-    build,
-    dist,
-    docs/
-    scripts/
-
-[isort]
-line_length = 99
-profile = black
-
-[tool:pytest]
-testpaths = tests
-generate_report_on_test = True
-markers =
-    multigpu
