+++ tests/conftest.py
@@ -142,7 +142,7 @@ def create_multiview_dataset_if_not_exists() -> None:
         dst_vid_top = os.path.join(dst_dir_vids, video.replace(".mp4", "_top.mp4"))
         dst_vid_bot = os.path.join(dst_dir_vids, video.replace(".mp4", "_bot.mp4"))
         ffmpeg_cmd = f"ffmpeg -i {src_vid} -filter_complex '[0]crop=iw:{y_split}:0:0[top];[0]crop=iw:ih-{y_split}:0:{y_split}[bot]' -map '[top]' {dst_vid_top} -map '[bot]' {dst_vid_bot}"  # noqa: E501
-        subprocess.run(ffmpeg_cmd, shell=True)
+        subprocess.run(ffmpeg_cmd, shell=True, check=True)
 
     # copy and split CollectedData.csv
     src_file = os.path.join(base_dir, "CollectedData.csv")
